+++ out/with_copyright.js
@@ -2,6 +2,6 @@
 /**
  * Copyright (c) 2023 TypeScript Company
  * Created: 2023-01-01 12:00:00
- * Last Updated: 2025-12-18 11:28
+ * Last Updated: 2025-12-18T11:44:24
  */
 //# sourceMappingURL=with_copyright.js.map