+++ test_files/old_timestamp.js
@@ -1,0 +1,16 @@
+/*
+ * Copyright (c) 2024 Bivex
+ *
+ * Author: Bivex
+ * Available for contact via email: support@b-b.top
+ * For up-to-date contact information:
+ * https://github.com/bivex
+ *
+ * Created: 2024-12-18 10:30:45
+ * Last Updated: 2025-12-18T11:43:26
+ *
+ * Licensed under the MIT License.
+ * Commercial licensing available upon request.
+ */
+
+console.log("Test file with old timestamp format");